chore(hogql): move parser diagnostic scripts into posthog/hogql/scripts/
The PBT diagnostic / corpus / bench / log-corpus scripts were `_`-prefixed
files under `posthog/hogql/test/` — the underscore there only existed to
dodge pytest collection. They aren't tests; they're CLI tools. Moving them
to `posthog/hogql/scripts/` (alongside `build_grammar_strategies.py`) lets
them drop the prefix and sit where scripts belong.

- `test/_pbt_diagnostic.py`      → `scripts/pbt_diagnostic.py`
- `test/_pbt_corpus.py`          → `scripts/pbt_corpus.py`
- `test/_log_corpus_diagnostic.py` → `scripts/log_corpus_diagnostic.py`
- `test/bench/parser_bench.py`   → `scripts/parser_bench.py` (empty `bench/` removed)

Dedup: the cross-script shared surface — the AST diff-path vocabulary
(`_node_type` / `_diff_path` / `_format_diff_path`), the `DivergenceShape`
machinery (`_ast_mismatch_shape` / `_shape_for`), error normalisation
(`_normalize_error` / `_reject_error`), and the `_probe_backend` sanity
check — moves into `scripts/_diagnostic_common.py`. Previously these lived
in `_pbt_diagnostic.py` and the other three scripts imported them from it,
coupling every script to the diagnostic's module. They now import from the
shared helper instead; `pbt_diagnostic.py` shrinks to the shrinker + main().
The leading underscore on `_diagnostic_common.py` marks it as a helper
module, not a CLI.

The auto-generated grammar strategies (`_generated_grammar_strategies.py`,
`_grammar_token_strategies.py`) stay in `test/` — they're fixtures for
`test_parser_grammar_pbt.py`, not script-only helpers.

`log_corpus_diagnostic.py`'s local dump dir moves with it:
`test/.local/` → `scripts/.local/` (.gitignore updated).
