+++ posthog/hogql/scripts/log_corpus_diagnostic.py
@@ -5,7 +5,7 @@ Pulls the last 7 days of HogQL queries that team_id=2 (PostHog's own
 internal team) ran against production ClickHouse, runs both the oracle
 and candidate parsers over each, and reports where they disagree.
 
-Complement to `_pbt_diagnostic.py`: that grinds randomly-generated
+Complement to `pbt_diagnostic.py`: that grinds randomly-generated
 grammar surface (what HogQL *permits*); this runs over real HogQL
 source text as the product emits it (what production *uses*) — insight
 queries, hand-rolled HogQL panels, warehouse view materialisation, and
@@ -18,7 +18,7 @@ whatever Cloud users trigger via the API. Both are worth running.
    DB whose name contains `OFFLINE` so the `system.query_log` scan goes
    to the background-workload cluster, not the live one.
 2. **Download a redacted dump** via `hogli metabase:query --format json`
-   to the gitignored `posthog/hogql/test/.local/hogql_log_corpus.json`.
+   to the gitignored `posthog/hogql/scripts/.local/hogql_log_corpus.json`.
    JSON not TSV — Metabase's TSV writer doesn't escape embedded
    newlines and HogQL queries are routinely multi-line. The embedded
    SQL applies a chain of `replaceRegexpAll` redaction passes (emails,
@@ -35,16 +35,16 @@ whatever Cloud users trigger via the API. Both are worth running.
 ## Usage
 
     # Default — auto-discover, download, run cpp-vs-python parity:
-    PYTHONPATH=. python posthog/hogql/test/_log_corpus_diagnostic.py
+    PYTHONPATH=. python posthog/hogql/scripts/log_corpus_diagnostic.py
 
     # Iterate on a candidate parser without re-pulling the corpus:
-    PYTHONPATH=. python posthog/hogql/test/_log_corpus_diagnostic.py \\
+    PYTHONPATH=. python posthog/hogql/scripts/log_corpus_diagnostic.py \\
         --skip-download \\
         --candidate rust-backtrack-json \\
         --write-failures /tmp/rust-backtrack-log-fails.sql
 
     # Override the dump path entirely:
-    PYTHONPATH=. python posthog/hogql/test/_log_corpus_diagnostic.py \\
+    PYTHONPATH=. python posthog/hogql/scripts/log_corpus_diagnostic.py \\
         --input /some/other/dump.json
 """
 
@@ -69,11 +69,11 @@ django.setup()
 from posthog.hogql import ast
 from posthog.hogql.errors import BaseHogQLError
 from posthog.hogql.parser import parse_select
-from posthog.hogql.test._pbt_diagnostic import _diff_path, _format_diff_path, _node_type
+from posthog.hogql.scripts._diagnostic_common import _diff_path, _format_diff_path, _node_type
 from posthog.hogql.visitor import clear_locations
 
 REPO_ROOT = Path(__file__).resolve().parents[3]
-DEFAULT_DUMP = REPO_ROOT / "posthog" / "hogql" / "test" / ".local" / "hogql_log_corpus.json"
+DEFAULT_DUMP = REPO_ROOT / "posthog" / "hogql" / "scripts" / ".local" / "hogql_log_corpus.json"
 
 
 # ---------------------------------------------------------------------------
@@ -364,7 +364,7 @@ def _write_failures(path: Path, failures: list[Failure]) -> None:
         f.write(f"-- hogql_log_corpus_failures.sql — {len(failures)} entries\n")
         f.write(f"-- ({breakdown})\n")
         f.write("-- Each block: occurrences count, failure kind + detail, then the query.\n")
-        f.write("-- Generated by posthog/hogql/test/_log_corpus_diagnostic.py\n\n")
+        f.write("-- Generated by posthog/hogql/scripts/log_corpus_diagnostic.py\n\n")
         for i, fa in enumerate(failures):
             f.write("-- " + "=" * 76 + "\n")
             f.write(f"-- [{i + 1}/{len(failures)}] seen {fa.n_occurrences}x in last 7d\n")
